Split the two files inital attempt
